Add: 엔티티 관계도, 게시판 api 유저 닉네임 표시

+++ src/user/entities/user.entity.ts
@@ -13,7 +13,7 @@ export class UserEntity {
 
     @Column({
         nullable: false,
-        type: 'tinyint'
+        type: 'smallint'
     })
     enrolled: number;
